refactor: move parsing logic from main.cpp to parsing.cpp
