+++ cgi-bin/storage/config_file.conf
@@ -11,13 +11,13 @@ server {
 	error_page 404 /error_page/404.html;
 	error_page 408 /error_page/408.html;
 	error_page 405 /error_page/405.html;
-	error_page 413 /error_page/413.html;
+	#error_page 413 /error_page/413.html;
 	error_page 500 /error_page/500.html;
 	error_page 501 /error_page/501.html;
 	error_page 504 /error_page/504.html;
 	error_page 505 /error_page/505.html;
 
-	client_body_size_buf 10000000;
+	#client_body_size_buf 100000000;
 
 	#test root/index
 	location / {
@@ -81,9 +81,9 @@ server {
 
 	location / {
 		root /home/yphang/Desktop/Webserv;
-		index html/upload.html;
+		index html/login.html;
 		autoindex on;
-		methods POST ;
+		methods POST GET;
 	}
 }
