+++ config_file.conf
@@ -8,16 +8,16 @@ server {
 	error_page 400 /error_page/400.html;
 	error_page 401 /error_page/401.html;
 	error_page 403 /error_page/403.html;
-	#error_page 404 /error_page/404.html;
+	error_page 404 /error_page/404.html;
 	error_page 408 /error_page/408.html;
 	error_page 405 /error_page/405.html;
-	error_page 413 /error_page/413.html;
+	#error_page 413 /error_page/413.html;
 	error_page 500 /error_page/500.html;
 	error_page 501 /error_page/501.html;
 	error_page 504 /error_page/504.html;
 	error_page 505 /error_page/505.html;
 
-	client_body_size_buf 10000000;
+	#client_body_size_buf 100000000;
 
 	#test root/index
 	location / {
@@ -29,6 +29,13 @@ server {
 	#test alias
 	location /test1 {
 		alias /home/yphang/Desktop/Webserv;
+		methods GET POST;
+	}
+
+	#test cookie
+	location /login {
+		alias /home/yphang/Desktop/Webserv;
+		index html/login.html;
 		methods GET POST;
 	}
 
@@ -74,21 +81,8 @@ server {
 	listen 127.0.0.1:2323;
 	#server_name server_1;
 
-	error_page 400 /error_page/400.html;
-	error_page 401 /error_page/401.html;
-	error_page 403 /error_page/403.html;
-	error_page 404 /error_page/404.html;
-	error_page 408 /error_page/408.html;
-	error_page 405 /error_page/405.html;
-	error_page 413 /error_page/413.html;
-	error_page 500 /error_page/500.html;
-	error_page 501 /error_page/501.html;
-	error_page 504 /error_page/504.html;
-	error_page 505 /error_page/505.html;
-
 	root /home/yphang/Desktop/Webserv;
 
-	error_page 404 /error_page/404.html;
 	client_body_size_buf 100000000;
 
 	location / {
@@ -97,6 +91,13 @@ server {
 		autoindex on;
 		methods POST GET;
 	}
+	
+	location /login {
+		alias /home/yphang/Desktop/Webserv;
+		index html/login.html;
+		methods GET POST;
+	}
+
 }
 
 server {
